debug

+++ os/services/rpl-border-router/embedded/slip-bridge.c
@@ -178,10 +178,10 @@ output(void)
        */
       
       // memcpy(UIP_IP_BUF[coap_packet_start_location + 8], &(tsch_current_asn.ls4b), 4)
-      PRINTF("\nStart_ASN_Numbers:%08x\n",startASN);
-      PRINTF("End_ASN_Numbers:%08x\n",tsch_current_asn.ls4b);
+      PRINTF("\nStart_ASN_Numbers:%08x\n",(unsigned int)startASN);
+      PRINTF("End_ASN_Numbers:%08x\n",(unsigned int)tsch_current_asn.ls4b);
       //PRINTF("Event_Counter:%d\n",eventCounter);
-      PRINTF("The_Packet_Latancy_is:%u",((tsch_current_asn.ls4b - startASN)) * 10 ); //ms time.
+      PRINTF("The_Packet_Latancy_is:%u",(unsigned int)(((tsch_current_asn.ls4b - startASN)) * 10)); //ms time.
       PRINTF(" ms\n");
       //PRINTF("The Packet Latancy is %u ms. \n",((tsch_current_asn.ls4b - startASN) - 4294967296) * 10 ); //ms time.                                       
       //PRINTF("Traffic_Classes:%02x. \n",UIP_IP_BUF->tcflow);
